sportswar/requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
billiard==3.3.0.23
celery==3.1.24
decorator==4.0.10
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.10.2
django-allauth==0.28.0
django-rest-auth==0.8.1
djangorestframework==3.4.7
gunicorn==19.6.0
ipython==5.1.0
ipython-genutils==0.1.0
kombu==3.0.37
oauthlib==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
python3-openid==3.0.10
pytz==2016.7
requests==2.11.1
requests-oauthlib==0.7.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
whitenoise==3.2
